requirements.txt
accelerate==1.0.1
annotated-types==0.7.0
appdirs==1.4.4
astor==0.8.1
# asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733175639022/work
attr==0.3.2
autograd==1.7.0
autograd-gamma==0.5.0
# backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
beartype==0.19.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
# comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.1.1
-e git+https://github.com/ibrahimethemhamamci/CT-CLIP.git@a00eceb016c7a1feb2256f0dde7c2b27fd7a6464#egg=ct_clip&subdirectory=CT_CLIP
cycler==0.12.1
# debugpy @ file:///croot/debugpy_1690905042057/work
# decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
docker-pycreds==0.4.0
einops==0.8.0
ema-pytorch==0.7.7
# et_xmlfile==2.0.0
eval_type_backport==0.2.2
# executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
# filelock==3.16.1
fonttools==4.55.3
formulaic==1.1.1
fsspec==2024.12.0
ftfy==6.2.0
gitdb==4.0.12
GitPython==3.1.44
graphlib_backport==1.1.0
h5py==3.11.0
huggingface-hub==0.27.0
idna==3.10
# imagenetv2_pytorch @ git+https://github.com/modestyachts/ImageNetV2_pytorch.git@14d4456c39fe7f02a665544dd9fc37c1a5f8b635
# imagenetv2_pytorch
# importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
importlib_resources==6.4.5
interface-meta==1.3.0
# ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
# ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1683289033986/work
# jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.5
joblib==1.4.2
# jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
# jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1727163409502/work
kiwisolver==1.4.7
lifelines==0.27.8
littleutils==0.2.4
MarkupSafe==2.1.5
matplotlib==3.7.5
# matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mpmath==1.3.0
# nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.1
nibabel==5.2.1
nltk==3.9.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
ogb==1.3.6
opencv-python==4.10.0.84
openpyxl==3.1.5
outdated==0.2.2
# packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.0.3
# parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
peft==0.13.2
# pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
# pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
# platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
# prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1727341649933/work
protobuf==5.29.3
# psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
# ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
# pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
pydantic==2.10.5
pydantic_core==2.27.2
# Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.4
# python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz==2024.2
PyYAML==6.0.2
# pyzmq @ file:///croot/pyzmq_1705605076900/work
regex==2024.11.6
requests==2.32.3
safetensors==0.5.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
# six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.2
# stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.13.3
torch==2.4.1
torchinfo==1.8.0
torchsummary==1.5.1
torchtyping==0.1.5
torchvision==0.19.1
# tornado @ file:///croot/tornado_1718740109488/work
tqdm==4.67.1
# traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
-e git+https://github.com/ibrahimethemhamamci/CT-CLIP.git@a00eceb016c7a1feb2256f0dde7c2b27fd7a6464#egg=transformer_maskgit&subdirectory=transformer_maskgit
transformers==4.30.1
triton==3.0.0
typeguard==2.13.3
# typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.2
urllib3==2.2.3
vector-quantize-pytorch==1.1.2
wandb==0.19.4
# wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
wilds==2.0.0
wrapt==1.17.0
XlsxWriter==3.2.0
zipp==3.20.2
